price-alert-script/requirements.txt
boto3
pylint
pytest
bs4
python-dotenv
requests
psycopg2-binary
